Checkpoint direct JS runtime support

+++ src/Tsonic.CSharp.Js/Date.cs
@@ -145,7 +145,7 @@ namespace Tsonic.CSharp.Js
         /// <summary>
         /// Returns current time in milliseconds since epoch
         /// </summary>
-        public static long now() => (long)(DateTimeOffset.UtcNow - Epoch).TotalMilliseconds;
+        public static double now() => (DateTimeOffset.UtcNow - Epoch).TotalMilliseconds;
 
         /// <summary>
         /// Parse date string and return milliseconds since epoch
@@ -183,7 +183,7 @@ namespace Tsonic.CSharp.Js
         /// <summary>
         /// Get milliseconds since epoch
         /// </summary>
-        public long getTime() => (long)(_value - Epoch).TotalMilliseconds;
+        public double getTime() => (_value - Epoch).TotalMilliseconds;
 
         /// <summary>
         /// Get full year (4 digits)
@@ -497,6 +497,6 @@ namespace Tsonic.CSharp.Js
         /// <summary>
         /// Get primitive value (milliseconds since epoch)
         /// </summary>
-        public long valueOf() => getTime();
+        public double valueOf() => getTime();
     }
 }
